Add Vagrantfile
Adds a Vagrantfile that allows running the dev instance in a VM.

Makes it easier to start developing and everyone's working on the same system, also no pollution of the main OS with utilities.

The provisioning sets everything up:

* Installs Node.js
* Sets the right permissions for global NPM modules
* Installs the required global modules
* Runs `npm install` and `webpack`

Ports 7070 and 8080 are forwarded to the main machine, so access through `localhost:7070` on the main system's browser is possible.

Step by step installation:

* Install [VirtualBox](https://www.virtualbox.org/wiki/Downloads) - *tested under VB 5.0*
* Install [Vagrant](https://www.vagrantup.com/downloads.html) - *tested under Vagrant 1.8.1*
* Check out or update current goontube repository version
* `cd` into working directory
* Run `vagrant up` - this will take a while, since it downloads the image, sets up the VM and installs everything
* `vagrant ssh` to connect to the machine
* Change into the web dir `cd /var/www` - this is a symlink to the repository itself
* Run `npm start`

Done. Hopefully. This was a pain to get running without a hitch.